ai/caption/requirements.txt
absl-py==1.3.0
alembic==1.10.2
asgiref==3.7.2
astor==0.8.1
astunparse==1.6.3
bcrypt==4.0.1
bleach==1.5.0
boto3==1.26.151
botocore==1.29.151
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.6
Django==4.2.4
filelock==3.9.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.5.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.25.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.51.1
h5py==3.10.0
html5lib==0.9999999
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kafka-python==2.0.2
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.2
ml-dtypes==0.2.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.0
pandas==2.1.4
pdf2image==1.16.3
Pillow==9.4.0
platformdirs==3.1.1
protobuf==3.20.3
psycopg2==2.9.1
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.6.0
PyMuPDF==1.22.3
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
quick-mailer==2022.2.22
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.13.0
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==2.0.6
sqlparse==0.4.4
stripe==5.4.0
tensorboard==1.8.0
tensorboard-data-server==0.7.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.8.0-py3-none-any.whl
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.21.0
Werkzeug==2.2.2
wrapt==1.14.1
